integrated somewhat of a permission system
#15
Es wird nun abgefragt ob die SteamID unter einer der Gruppen hinterlegt ist. Entsprechend dem werden bestimmte Zeiten und Funktionen freigegeben.

+++ assets/components/comdelete.php
@@ -9,12 +9,20 @@
 <body>
 
 <?php
+require '../steamauth/steamauth.php';
 include("fb37dbconnect.php");
+include ('../steamauth/userInfo.php');
+include("fb37allowedids.php");
+
 $dbconnect=mysqli_connect($hostname,$username,$password,$dbname);
 $id=$_REQUEST['id'];
 $mid=$_REQUEST['mid'];
+if (in_array($steamprofile['steamid'], $admin)) {
 $result = mysqli_query($dbconnect,"DELETE FROM memberComments WHERE id='".$id."'") or die ( mysqli_error());
 header("Location: ../../assets/components/mitarbeiterprofil.php?id=".$mid);
+} else {
+    header("Location: ../../assets/components/mitarbeiterprofil.php?id=".$mid);
+}
 ?>
